Rename compressed_tensor folder to compressed_tensors Move RunCompressedTest to the same file Update tests to unittest
